Clustering added
