+++ .env.example
@@ -5,3 +5,5 @@ db_user="devUser"
 db_pwd="PE7nojPn55MrhI4H"
 db_host="cluster0.fpvqy.mongodb.net"
 db_name="myFirstDatabase"
+
+CLUSTER=yes
